+++ src/Akeneo/Pim/WorkOrganization/TeamworkAssistant/Component/Exception/CannotRemoveUserGroupLinkedToProjectException.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Akeneo PIM Enterprise Edition.
+ *
+ * (c) 2021 Akeneo SAS (http://www.akeneo.com)
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Akeneo\Pim\WorkOrganization\TeamworkAssistant\Component\Exception;
+
+/**
+ * @author    Nicolas Marniesse <nicolas.marniesse@akeneo.com>
+ * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
+ */
+final class CannotRemoveUserGroupLinkedToProjectException extends \Exception
+{
+}