+++ CHANGELOG.md
@@ -43,6 +43,8 @@
 - PIM-9822: Fix Error 500 after deleting a filterable asset family attribute
 - PIM-9825: Fix Assets not being properly reindexed after bulk edit
 - PIM-9835: Fix DQI loading in attribute edit page with a large amount of options 
+- PIM-9854: Fix message when deleting a user group used in a project
+- PIM-9853: Make the word "product" translatable 
 
 ## Improvements